refactor: move the Parquet and XLSX writers out of filesql.go
filesql.go was 1,220 lines holding five things: the four exported entry points
-- Open, OpenContext, LoadInto and DumpDatabase -- the dump machinery they
share, and three writers, one per output shape.

The two that write a typed container format move to dump_parquet.go and
dump_xlsx.go, 281 and 270 lines. What is left in filesql.go is 687: the entry
points, the shared machinery, and the 125 lines that write the three text
formats, which stay because picking between them is what the dump machinery
does.

Not a line of the moved code changed. Concatenating the three files' bodies
back together reproduces the original 43,799 bytes exactly; the only difference
is that each file now names the imports it uses.
