+++ filesql.go
@@ -13,12 +13,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/nao1215/filesql/internal/infer"
 	"github.com/nao1215/filesql/internal/reader"
 	"github.com/nao1215/filesql/internal/writer"
 	"github.com/nao1215/filesql/parser"
-	"github.com/parquet-go/parquet-go"
-	"github.com/xuri/excelize/v2"
 )
 
 // Open creates an SQL database from CSV, TSV, or LTSV files.
@@ -687,534 +684,4 @@ func formatDumpValue(value any) string {
 	default:
 		return fmt.Sprintf("%v", v)
 	}
-}
-
-// writeParquetTableData writes SQLite table data to Parquet format
-func writeParquetTableData(w io.Writer, columns []string, rows *sql.Rows) error {
-	if len(columns) == 0 {
-		return fmt.Errorf("%w: no columns defined", ErrEmptyData)
-	}
-
-	// The declared types are read before the scan loop. Draining Rows closes it,
-	// and ColumnTypes on a closed Rows fails, which would leave a table with no
-	// rows with nothing to take its schema from.
-	declared := make([]string, len(columns))
-	if types, err := rows.ColumnTypes(); err == nil {
-		for i, ct := range types {
-			if i < len(declared) {
-				declared[i] = ct.DatabaseTypeName()
-			}
-		}
-	}
-
-	// Read all rows into memory first. The raw driver values are kept rather than
-	// their rendered text because Parquet is a typed format: which Parquet type
-	// each column is written as is decided from the values themselves, below.
-	var allRows [][]any
-
-	// Prepare for scanning
-	values := make([]any, len(columns))
-	scanArgs := make([]any, len(columns))
-	for i := range values {
-		scanArgs[i] = &values[i]
-	}
-
-	for rows.Next() {
-		if err := rows.Scan(scanArgs...); err != nil {
-			return fmt.Errorf("%w: failed to scan row: %w", ErrDatabaseOperation, err)
-		}
-		row := make([]any, len(columns))
-		copy(row, values)
-		allRows = append(allRows, row)
-	}
-
-	if err := rows.Err(); err != nil {
-		return fmt.Errorf("%w: error iterating rows: %w", ErrDatabaseOperation, err)
-	}
-
-	return writeParquetData(w, columns, allRows, declared)
-}
-
-// parquetKind is the Parquet type one column is written as.
-type parquetKind int
-
-const (
-	parquetString parquetKind = iota
-	parquetInt
-	parquetFloat
-)
-
-// parquetColumnKind decides how one column is written. A column is numeric only
-// when every value in it is, so a column that mixes a number with text is
-// written as STRING rather than losing the text: SQLite types values, not
-// columns, and a dump has to carry back whatever the rows actually held.
-//
-// The declared type decides an empty column, and only an empty one. It is what a
-// table emptied by the session still knows about itself, so an auto-save keeps
-// the schema instead of rewriting every column as STRING; for a column with rows
-// the values are the better witness, because a query's result columns carry no
-// declared type at all.
-func parquetColumnKind(rows [][]any, col int, declaredType string) parquetKind {
-	kind := parquetKind(-1)
-	lossy := false // an int64 in this column that float64 cannot carry back
-	for _, row := range rows {
-		// A blank cell says nothing about the column's type. SQLite stores a
-		// blank in a numeric column as the empty string, since "" has no numeric
-		// value to convert to, and letting that decide the column wrote a column
-		// of numbers as text the moment one row was missing an entry.
-		if col >= len(row) || row[col] == nil || row[col] == "" {
-			continue
-		}
-		var cell parquetKind
-		switch v := row[col].(type) {
-		case int64:
-			cell = parquetInt
-			lossy = lossy || !infer.Int64SurvivesFloat64(v)
-		case float64:
-			cell = parquetFloat
-		default:
-			return parquetString
-		}
-		switch {
-		case kind < 0:
-			kind = cell
-		case kind != cell:
-			// int64 and float64 in one column widen to float64, which holds
-			// both while every integer survives a float64 round-trip; one that
-			// does not makes the column STRING below, the answer a
-			// number-beside-text column already gets, because a dump has to
-			// carry back the value and DOUBLE would write a different number.
-			kind = parquetFloat
-		}
-	}
-	if kind == parquetFloat && lossy {
-		return parquetString
-	}
-	if kind >= 0 {
-		return kind
-	}
-	switch strings.ToUpper(declaredType) {
-	case "INTEGER", "INT", "BIGINT":
-		return parquetInt
-	case "REAL", "FLOAT", "DOUBLE":
-		return parquetFloat
-	default:
-		return parquetString
-	}
-}
-
-// parquetNodeFor is the Parquet schema node a parquetKind is written as. Every
-// column is optional, so a stored null survives the round-trip.
-func parquetNodeFor(kind parquetKind) parquet.Node {
-	switch kind {
-	case parquetInt:
-		return parquet.Optional(parquet.Leaf(parquet.Int64Type))
-	case parquetFloat:
-		return parquet.Optional(parquet.Leaf(parquet.DoubleType))
-	default:
-		return parquet.Optional(parquet.String())
-	}
-}
-
-// orderedGroup is parquet.Group with its fields kept in the order the table
-// declares its columns; parquet.Group alone sorts them by name, and a dump
-// that reordered its columns would hand every downstream reader a different
-// table than the one it was asked to write.
-type orderedGroup struct {
-	parquet.Group
-	names []string
-}
-
-// Fields returns the group's fields in declaration order.
-func (g orderedGroup) Fields() []parquet.Field {
-	sorted := g.Group.Fields()
-	byName := make(map[string]parquet.Field, len(sorted))
-	for _, field := range sorted {
-		byName[field.Name()] = field
-	}
-	fields := make([]parquet.Field, 0, len(g.names))
-	for _, name := range g.names {
-		fields = append(fields, byName[name])
-	}
-	return fields
-}
-
-// parquetCellValue renders one cell as the column's Parquet value. ok is false
-// for a null: a nil value is the SQL NULL the row carried, and a blank in a
-// numeric column is written as the null it means, since a number has no
-// spelling for a blank. A text column keeps its empty string. A value that does
-// not match a numeric column's kind cannot occur, because parquetColumnKind
-// only chooses a numeric kind when every value in the column is numeric.
-func parquetCellValue(kind parquetKind, value any) (parquet.Value, bool, error) {
-	if value == nil {
-		return parquet.Value{}, false, nil
-	}
-	if text, isText := value.(string); isText && text == "" && kind != parquetString {
-		return parquet.Value{}, false, nil
-	}
-	switch kind {
-	case parquetInt:
-		n, ok := value.(int64)
-		if !ok {
-			return parquet.Value{}, false, fmt.Errorf("%w: %T in an integer column", ErrInvalidData, value)
-		}
-		return parquet.Int64Value(n), true, nil
-	case parquetFloat:
-		switch v := value.(type) {
-		case float64:
-			return parquet.DoubleValue(v), true, nil
-		case int64:
-			return parquet.DoubleValue(float64(v)), true, nil
-		default:
-			return parquet.Value{}, false, fmt.Errorf("%w: %T in a real column", ErrInvalidData, value)
-		}
-	default:
-		return parquet.ByteArrayValue([]byte(formatDumpValue(value))), true, nil
-	}
-}
-
-// writeParquetData writes data to Parquet format. A nil cell in rows is stored
-// as a Parquet null, so a SQL NULL survives the round-trip instead of collapsing
-// into an empty string. declared holds each column's declared SQL type, which
-// decides a column that has no rows to speak for it.
-//
-// Each column is written as the Parquet type its values call for rather than as
-// STRING. Parquet states the type of every column in its schema and readers
-// trust it, so writing a numeric column as digit strings hands the next tool a
-// column it will compare and sort as text.
-//
-// A table with no rows is written as a schema with no row groups, which is a
-// valid Parquet file: the other formats write their header and nothing else, and
-// a dump that refused to write an emptied table let an auto-save keep the rows
-// the caller had deleted.
-func writeParquetData(w io.Writer, columns []string, rows [][]any, declared []string) error {
-	if len(columns) == 0 {
-		return fmt.Errorf("%w: no columns defined", ErrEmptyData)
-	}
-
-	kinds := make([]parquetKind, len(columns))
-	group := orderedGroup{Group: make(parquet.Group, len(columns)), names: columns}
-	for i, col := range columns {
-		declaredType := ""
-		if i < len(declared) {
-			declaredType = declared[i]
-		}
-		kinds[i] = parquetColumnKind(rows, i, declaredType)
-		group.Group[col] = parquetNodeFor(kinds[i])
-	}
-	schema := parquet.NewSchema("table", group)
-
-	// The writer takes an io.Writer and never closes it, so the caller keeps
-	// ownership of the destination.
-	writer := parquet.NewGenericWriter[any](w, schema)
-
-	// Rows are built as Parquet values directly rather than deconstructed from
-	// Go values by reflection, which reads a zero value in an optional column
-	// as a null and would turn a stored 0 or "" into a missing cell.
-	buf := make([]parquet.Row, 0, min(len(rows), 1024))
-	flush := func() error {
-		if len(buf) == 0 {
-			return nil
-		}
-		if _, err := writer.WriteRows(buf); err != nil {
-			return fmt.Errorf("%w: failed to write rows to parquet: %w", ErrIOOperation, err)
-		}
-		buf = buf[:0]
-		return nil
-	}
-	for _, row := range rows {
-		out := make(parquet.Row, len(columns))
-		for i := range columns {
-			var value any
-			if i < len(row) {
-				value = row[i]
-			}
-			cell, present, err := parquetCellValue(kinds[i], value)
-			if err != nil {
-				return err
-			}
-			if present {
-				out[i] = cell.Level(0, 1, i)
-			} else {
-				out[i] = parquet.NullValue().Level(0, 0, i)
-			}
-		}
-		buf = append(buf, out)
-		if len(buf) == cap(buf) {
-			if err := flush(); err != nil {
-				return err
-			}
-		}
-	}
-	if err := flush(); err != nil {
-		return err
-	}
-
-	// Close writes the footer, so this is where an incomplete file shows up.
-	// For a table with no rows it writes the schema and the footer, which is
-	// the whole file.
-	if err := writer.Close(); err != nil {
-		return fmt.Errorf("%w: failed to close parquet writer: %w", ErrIOOperation, err)
-	}
-	return nil
-}
-
-// xlsxSheet is one sheet of a workbook being written. Rows are opened when the
-// sheet is reached rather than up front, because a *sql.Rows holds a cursor and
-// only one can be read at a time.
-type xlsxSheet struct {
-	// name is the sheet name, already adapted to what Excel accepts.
-	name string
-	// open yields the sheet's columns and rows.
-	open func() ([]string, *sql.Rows, error)
-}
-
-// writeXLSXTableData writes SQLite table data to Excel XLSX format as a
-// single-sheet workbook named after the table.
-func writeXLSXTableData(w io.Writer, tableName string, columns []string, rows *sql.Rows) error {
-	// The sheet name is what a reader turns back into a table name, so it comes
-	// from the table this dump is of, adapted to what Excel accepts.
-	return writeXLSXWorkbook(w, []xlsxSheet{{
-		name: excelSheetName(tableName),
-		open: func() ([]string, *sql.Rows, error) { return columns, rows, nil },
-	}})
-}
-
-// writeXLSXWorkbook writes sheets as one workbook. A workbook overwritten in
-// place goes through here with every one of its sheets, so a file of several
-// sheets comes back whole rather than being refused or flattened to one.
-func writeXLSXWorkbook(w io.Writer, sheets []xlsxSheet) error {
-	return writeXLSXWorkbookOnto(w, nil, sheets)
-}
-
-// writeXLSXWorkbookOnto writes sheets into base, or into a new workbook when
-// base is nil.
-//
-// A save that replaces a workbook writes onto the workbook it is replacing, so
-// what this package does not hold survives the save: a sheet the sheet policy
-// chose not to load used to be deleted from the caller's file, and a column
-// width, a merged range and a comment were gone from the sheets it did load.
-// Only the rows of a sheet a table was loaded from are rewritten, since those
-// rows are what the table is.
-func writeXLSXWorkbookOnto(w io.Writer, base *excelize.File, sheets []xlsxSheet) error {
-	if len(sheets) == 0 {
-		return fmt.Errorf("%w: no sheets to write", ErrEmptyData)
-	}
-
-	f := base
-	if f == nil {
-		f = excelize.NewFile()
-	}
-	defer func() {
-		_ = f.Close() // Ignore close error
-	}()
-
-	for _, sheet := range sheets {
-		had, err := xlsxSheetBefore(f, sheet.name, base != nil)
-		if err != nil {
-			return err
-		}
-		written, err := writeXLSXSheet(f, sheet, had.values)
-		if err != nil {
-			return err
-		}
-		if err := trimXLSXSheet(f, sheet.name, had.extent, written); err != nil {
-			return err
-		}
-	}
-
-	// excelize starts a workbook with a default sheet. It is only ours to remove
-	// once a sheet of our own exists, and not at all if a sheet reused its name,
-	// and never when the workbook came from the caller rather than from here.
-	if _, err := f.GetSheetIndex(defaultSheetName); err == nil && base == nil {
-		hasOwn := false
-		for _, sheet := range sheets {
-			if sheet.name == defaultSheetName {
-				hasOwn = true
-				break
-			}
-		}
-		if !hasOwn {
-			if err := f.DeleteSheet(defaultSheetName); err != nil {
-				return fmt.Errorf("failed to delete default sheet: %w", err)
-			}
-		}
-	}
-
-	// Why Write and not SaveAs: SaveAs picks the container format from the file
-	// extension, and the caller stages the write, so the only name available here
-	// carries a temporary suffix that Excel rejects. Any compression the caller
-	// asked for is already wrapped around w.
-	if err := f.Write(w); err != nil {
-		return fmt.Errorf("%w: failed to write Excel file: %w", ErrIOOperation, err)
-	}
-
-	return nil
-}
-
-// unchangedXLSXCell reports whether the cell at row and column, both counted
-// from one, already reads as value in the sheet a save is writing onto.
-//
-// A cell holds more than the text a table can carry. It may hold a formula and
-// the value that formula last evaluated to, and it may hold a date as a serial
-// number with a format that renders it. Writing the loaded text back over such
-// a cell says nothing the cell did not already say and takes the rest with it:
-// the formula became an empty cell, so the workbook lost the rule that produced
-// its numbers, and the date became text, so the column no longer sorted or
-// calculated as dates. Leaving the cell alone keeps both. A cell whose value did
-// change is written, which is the point of the save, and a formula that no
-// longer produces the value it is next to cannot stay.
-func unchangedXLSXCell(before [][]string, row, column int, value string) bool {
-	if row > len(before) {
-		return false
-	}
-	cells := before[row-1]
-	if column > len(cells) {
-		// A workbook stores no cell for a trailing empty one, so a row that ends
-		// before this column holds nothing here, and nothing is what an empty
-		// value would write.
-		return value == ""
-	}
-	return cells[column-1] == value
-}
-
-// xlsxExtent is how far a sheet's values reached before it was rewritten.
-type xlsxExtent struct {
-	rows    int
-	columns int
-}
-
-// xlsxSheetPrior is a sheet as it stood before a save wrote onto it.
-type xlsxSheetPrior struct {
-	// extent is how far its values reached, so what the table no longer covers
-	// can be removed afterwards.
-	extent xlsxExtent
-	// values is the sheet as the loader read it, which is what a cell has to be
-	// compared against to tell an edit from a value that came out of the file
-	// unchanged. It is nil for a workbook being built from nothing.
-	values [][]string
-}
-
-// xlsxSheetBefore reads a sheet about to be rewritten. Nothing is removed up
-// front: writing over a cell keeps the style it carries, where clearing the
-// sheet first would take the styles, the merged ranges and the comments with
-// it. A sheet the workbook does not have yet has nothing to read, and neither
-// does a workbook this package is building from nothing.
-//
-// The values are read the way the loader read them, dates included, so a cell
-// that comes back the same string it went in as is recognizable as untouched.
-func xlsxSheetBefore(f *excelize.File, sheetName string, ontoExisting bool) (xlsxSheetPrior, error) {
-	if !ontoExisting {
-		return xlsxSheetPrior{}, nil
-	}
-	if _, err := f.GetSheetIndex(sheetName); err != nil {
-		return xlsxSheetPrior{}, nil //nolint:nilerr // A sheet that is not there yet; writeXLSXSheet creates it.
-	}
-	rows, err := f.GetRows(sheetName)
-	if err != nil {
-		return xlsxSheetPrior{}, fmt.Errorf("failed to read sheet %s: %w", sheetName, err)
-	}
-	prior := xlsxSheetPrior{
-		extent: xlsxExtent{rows: len(rows)},
-		values: reader.NormalizeXLSXDates(f, sheetName, rows),
-	}
-	for _, row := range rows {
-		prior.extent.columns = max(prior.extent.columns, len(row))
-	}
-	return prior, nil
-}
-
-// trimXLSXSheet removes what the rewritten table no longer covers: rows below
-// its last one and columns to the right of its last. Removing from the far end
-// inward keeps the indexes gathered before the write correct as the sheet
-// shrinks. A sheet that grew has nothing to trim.
-func trimXLSXSheet(f *excelize.File, sheetName string, had xlsxExtent, wrote xlsxExtent) error {
-	for row := had.rows; row > wrote.rows; row-- {
-		if err := f.RemoveRow(sheetName, row); err != nil {
-			return fmt.Errorf("failed to remove row %d of sheet %s: %w", row, sheetName, err)
-		}
-	}
-	for column := had.columns; column > wrote.columns; column-- {
-		name, err := excelize.ColumnNumberToName(column)
-		if err != nil {
-			return fmt.Errorf("failed to name column %d of sheet %s: %w", column, sheetName, err)
-		}
-		if err := f.RemoveCol(sheetName, name); err != nil {
-			return fmt.Errorf("failed to remove column %s of sheet %s: %w", name, sheetName, err)
-		}
-	}
-	return nil
-}
-
-// writeXLSXSheet adds one sheet to f and fills it. A cell whose value matches
-// what before already holds is left alone.
-func writeXLSXSheet(f *excelize.File, sheet xlsxSheet, before [][]string) (xlsxExtent, error) {
-	columns, rows, err := sheet.open()
-	if err != nil {
-		return xlsxExtent{}, err
-	}
-	if rows != nil {
-		defer rows.Close()
-	}
-	if len(columns) == 0 {
-		return xlsxExtent{}, fmt.Errorf("%w: no columns defined", ErrEmptyData)
-	}
-
-	if sheet.name != defaultSheetName {
-		if _, err := f.NewSheet(sheet.name); err != nil {
-			return xlsxExtent{}, fmt.Errorf("failed to create sheet %s: %w", sheet.name, err)
-		}
-	}
-
-	// Set headers
-	for i, col := range columns {
-		if unchangedXLSXCell(before, 1, i+1, col) {
-			continue
-		}
-		cell, err := excelize.CoordinatesToCellName(i+1, 1)
-		if err != nil {
-			return xlsxExtent{}, fmt.Errorf("failed to generate cell name for column %d: %w", i+1, err)
-		}
-		if err := f.SetCellValue(sheet.name, cell, col); err != nil {
-			return xlsxExtent{}, fmt.Errorf("failed to set header %s: %w", col, err)
-		}
-	}
-
-	// Prepare for scanning rows
-	values := make([]interface{}, len(columns))
-	scanArgs := make([]interface{}, len(columns))
-	for i := range values {
-		scanArgs[i] = &values[i]
-	}
-
-	// Write data rows
-	rowIndex := 2 // Start from row 2 (after header)
-	for rows.Next() {
-		if err := rows.Scan(scanArgs...); err != nil {
-			return xlsxExtent{}, fmt.Errorf("failed to scan row: %w", err)
-		}
-
-		for i, val := range values {
-			// Every cell is written as text, the same string the text formats
-			// produce, so one table dumped twice does not disagree with itself.
-			cellValue := formatDumpValue(val)
-			if unchangedXLSXCell(before, rowIndex, i+1, cellValue) {
-				continue
-			}
-			cell, err := excelize.CoordinatesToCellName(i+1, rowIndex)
-			if err != nil {
-				return xlsxExtent{}, fmt.Errorf("failed to generate cell name for column %d, row %d: %w", i+1, rowIndex, err)
-			}
-			if err := f.SetCellValue(sheet.name, cell, cellValue); err != nil {
-				return xlsxExtent{}, fmt.Errorf("failed to set cell value at %s: %w", cell, err)
-			}
-		}
-		rowIndex++
-	}
-
-	if err := rows.Err(); err != nil {
-		return xlsxExtent{}, fmt.Errorf("error reading rows: %w", err)
-	}
-
-	return xlsxExtent{rows: rowIndex - 1, columns: len(columns)}, nil
 }
